api_yamdb/api/mixins.py
from rest_framework import mixins, viewsets


class DestroyListCreatMixinSet(
    mixins.CreateModelMixin,
    mixins.ListModelMixin,
    mixins.DestroyModelMixin,
    viewsets.GenericViewSet
):
    pass
